Minor model fix

+++ src/main/java/ajm/spapp/api/model/AccessRequest.java
@@ -5,7 +5,6 @@ import lombok.*;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.Optional;
 
 @Entity
 @Table(name="AccessRequest")
@@ -18,7 +17,7 @@ public class AccessRequest {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "EmployeeId", nullable = false)
